refactor(Verify): Move validate helper function to its own module
#21
